backend/requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
pydantic>=2.0.0,<3.0.0
pydantic[email]
pydantic-settings>=2.0.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.0,<4.1.0
psycopg2-binary>=2.9.0,<3.0.0
httpx>=0.27.0,<1.0.0
nltk>=3.8.0,<4.0.0
stripe>=8.0.0,<10.0.0
python-multipart>=0.0.9,<1.0.0
python-dotenv>=1.0.0,<2.0.0
